Rename new cache classes
